Code cleanup
Clean up the code for copying policies and hub templating workaround
