+++ bundle/manifests/cluster-group-upgrades-operator.clusterserviceversion.yaml
@@ -134,8 +134,6 @@ spec:
         path: computedMaxConcurrency
       - displayName: Conditions
         path: conditions
-      - displayName: Copied Policies
-        path: copiedPolicies
       - displayName: Managed Policies Compliant Before Upgrade
         path: managedPoliciesCompliantBeforeUpgrade
       - displayName: Managed Policies Content
