+++ controllers/clustergroupupgrade_controller.go
@@ -1096,195 +1096,6 @@ func (r *ClusterGroupUpgradeReconciler) doManagedPoliciesExist(
 	return true, managedPoliciesInfo, nil
 }
 
-func (r *ClusterGroupUpgradeReconciler) copyManagedInformPolicy(
-	ctx context.Context, clusterGroupUpgrade *ranv1alpha1.ClusterGroupUpgrade, managedPolicy *unstructured.Unstructured) (string, error) {
-
-	// Create a new unstructured variable to keep all the information for the new policy.
-	newPolicy := &unstructured.Unstructured{}
-
-	// Set new policy name, namespace, group, kind and version.
-	name := utils.GetResourceName(clusterGroupUpgrade, managedPolicy.GetName())
-	newPolicy.SetName(name)
-	newPolicy.SetNamespace(clusterGroupUpgrade.GetNamespace())
-	newPolicy.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "policy.open-cluster-management.io",
-		Kind:    "Policy",
-		Version: "v1",
-	})
-
-	// Set new policy labels.
-	labels := managedPolicy.GetLabels()
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	labels["app"] = "openshift-cluster-group-upgrades"
-	labels["openshift-cluster-group-upgrades/clusterGroupUpgrade"] = clusterGroupUpgrade.Name
-	labels["openshift-cluster-group-upgrades/parentPolicyName"] = managedPolicy.GetName()
-	labels[utils.ExcludeFromClusterBackup] = "true"
-	newPolicy.SetLabels(labels)
-
-	// Set new policy annotations - copy them from the managed policy.
-	annotations := managedPolicy.GetAnnotations()
-	if annotations == nil {
-		annotations = make(map[string]string)
-	}
-	annotations[utils.DesiredResourceName] = utils.PrefixNameWithNamespace(managedPolicy.GetNamespace(), name)
-	newPolicy.SetAnnotations(annotations)
-
-	// Set new policy remediationAction.
-	newPolicy.Object["spec"] = managedPolicy.Object["spec"]
-	specObject := newPolicy.Object["spec"].(map[string]interface{})
-	specObject["remediationAction"] = utils.RemediationActionEnforce
-
-	// Update the ConfigurationPolicy of the new policy.
-	err := r.updateConfigurationPolicyForCopiedPolicy(ctx, clusterGroupUpgrade, newPolicy, managedPolicy.GetName(), managedPolicy.GetNamespace())
-	if err != nil {
-		return "", err
-	}
-
-	// Create the new policy in the desired namespace.
-	err = r.createNewPolicyFromStructure(ctx, clusterGroupUpgrade, newPolicy)
-	if err != nil {
-		r.Log.Info("Error creating policy", "err", err)
-		return "", err
-	}
-	return newPolicy.GetName(), nil
-}
-
-func (r *ClusterGroupUpgradeReconciler) updateConfigurationPolicyForCopiedPolicy(
-	ctx context.Context, clusterGroupUpgrade *ranv1alpha1.ClusterGroupUpgrade, policy *unstructured.Unstructured, managedPolicyName, managedPolicyNamespace string) error {
-
-	// Go through the policy policy-templates.
-	policySpec := policy.Object["spec"].(map[string]interface{})
-	policyTemplates := policySpec["policy-templates"].([]interface{})
-	for _, plcTmpl := range policyTemplates {
-		// Update the metadata name of the ConfigurationPolicy.
-		plcTmplDef := plcTmpl.(map[string]interface{})["objectDefinition"].(map[string]interface{})
-		metadata := plcTmplDef["metadata"]
-		r.updateConfigurationPolicyName(clusterGroupUpgrade, metadata)
-
-		// Ensure the resources referenced in the hub template policy exist if applicable
-		plcTmplDefSpec := plcTmplDef["spec"].(map[string]interface{})
-
-		// One and only one of [object-templates, object-templates-raw] should be defined
-		objectTemplatePresent := plcTmplDefSpec[utils.ObjectTemplates] != nil
-		objectTemplateRawPresent := plcTmplDefSpec[utils.ObjectTemplatesRaw] != nil
-
-		var configPlcTmpls interface{}
-
-		switch {
-		case objectTemplatePresent && objectTemplateRawPresent:
-			return fmt.Errorf("[updateConfigurationPolicyForCopiedPolicy] found both %s and %s in policyTemplate", utils.ObjectTemplates, utils.ObjectTemplatesRaw)
-		case !objectTemplatePresent && !objectTemplateRawPresent:
-			return fmt.Errorf("[updateConfigurationPolicyForCopiedPolicy] can't find %s or %s in policyTemplate", utils.ObjectTemplates, utils.ObjectTemplatesRaw)
-		case objectTemplatePresent:
-			configPlcTmpls = plcTmplDefSpec[utils.ObjectTemplates].([]interface{})
-
-			resolvedconfigPlcTmpls, err := r.updateConfigurationPolicyHubTemplate(
-				ctx, configPlcTmpls, clusterGroupUpgrade.GetNamespace(), managedPolicyName, managedPolicyNamespace)
-			if err != nil {
-				return err
-			}
-
-			plcTmplDefSpec[utils.ObjectTemplates] = resolvedconfigPlcTmpls
-
-		case objectTemplateRawPresent:
-			// We cannot strip the raw templating here because they must be passed down to ACM
-			configPlcTmpls := plcTmplDefSpec[utils.ObjectTemplatesRaw].(string)
-
-			resolvedconfigPlcTmpls, err := r.updateConfigurationPolicyHubTemplate(
-				ctx, configPlcTmpls, clusterGroupUpgrade.GetNamespace(), managedPolicyName, managedPolicyNamespace)
-			if err != nil {
-				return err
-			}
-
-			stringResolved, err := utils.YamlToString(resolvedconfigPlcTmpls)
-			if err != nil {
-				return err
-			}
-
-			// If we do not trim the vertical bar here then the policy will be created with two
-			// vertical bars and will not be able to be marshalled later
-			plcTmplDefSpec[utils.ObjectTemplatesRaw] = strings.TrimPrefix(stringResolved, "|")
-		default:
-			return fmt.Errorf("[updateConfigurationPolicyForCopiedPolicy] can't find %s or %s in policyTemplate", utils.ObjectTemplates, utils.ObjectTemplatesRaw)
-		}
-	}
-
-	return nil
-}
-
-func (r *ClusterGroupUpgradeReconciler) updateConfigurationPolicyHubTemplate(
-	ctx context.Context, objectTmpl interface{}, cguNamespace, managedPolicyName, managedPolicyNamespace string) (interface{}, error) {
-	// Process only if the managed policy is not created in the CGU namespace
-	if managedPolicyNamespace == cguNamespace {
-		return objectTmpl, nil
-	}
-
-	tmplResolver := &utils.TemplateResolver{
-		Client:          r.Client,
-		Ctx:             ctx,
-		TargetNamespace: cguNamespace,
-		PolicyName:      managedPolicyName,
-		PolicyNamespace: managedPolicyNamespace,
-	}
-
-	resolvedObjectTmpl, err := tmplResolver.ProcessHubTemplateFunctions(objectTmpl)
-	if err != nil {
-		return resolvedObjectTmpl, err
-	}
-
-	return resolvedObjectTmpl, nil
-}
-
-func (r *ClusterGroupUpgradeReconciler) updateConfigurationPolicyName(
-	clusterGroupUpgrade *ranv1alpha1.ClusterGroupUpgrade, metadata interface{}) {
-
-	metadataContent := metadata.(map[string]interface{})
-	name := utils.GetResourceName(clusterGroupUpgrade, metadataContent["name"].(string))
-	safeName := utils.GetSafeResourceName(name, "", clusterGroupUpgrade, utils.MaxObjectNameLength)
-	metadataContent["name"] = safeName
-}
-
-func (r *ClusterGroupUpgradeReconciler) createNewPolicyFromStructure(
-	ctx context.Context, clusterGroupUpgrade *ranv1alpha1.ClusterGroupUpgrade, policy *unstructured.Unstructured) error {
-
-	name := policy.GetName()
-	safeName := utils.GetSafeResourceName(name, policy.GetNamespace(), clusterGroupUpgrade, utils.MaxPolicyNameLengthExcludingTheDot)
-	policy.SetName(safeName)
-	if err := controllerutil.SetControllerReference(clusterGroupUpgrade, policy, r.Scheme); err != nil {
-		return err
-	}
-	existingPolicy := &unstructured.Unstructured{}
-	existingPolicy.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "policy.open-cluster-management.io",
-		Kind:    "Policy",
-		Version: "v1",
-	})
-	err := r.Client.Get(ctx, client.ObjectKey{
-		Name:      safeName,
-		Namespace: clusterGroupUpgrade.Namespace,
-	}, existingPolicy)
-
-	if err != nil {
-		if errors.IsNotFound(err) {
-			err = r.Client.Create(ctx, policy)
-			if err != nil {
-				return err
-			}
-		} else {
-			return err
-		}
-	} else {
-		policy.SetResourceVersion(existingPolicy.GetResourceVersion())
-		err = r.Client.Update(ctx, policy)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 func (r *ClusterGroupUpgradeReconciler) ensureBatchPlacementRule(ctx context.Context, clusterGroupUpgrade *ranv1alpha1.ClusterGroupUpgrade, managedPolicy *unstructured.Unstructured) (string, error) {
 
 	name := utils.GetResourceName(clusterGroupUpgrade, managedPolicy.GetName()+"-placement")
@@ -1569,25 +1380,6 @@ func (r *ClusterGroupUpgradeReconciler) getPlacementBindings(ctx context.Context
 	}
 
 	return placementBindingsList, nil
-}
-
-func (r *ClusterGroupUpgradeReconciler) getCopiedPolicies(ctx context.Context, clusterGroupUpgrade *ranv1alpha1.ClusterGroupUpgrade) (*unstructured.UnstructuredList, error) {
-	var policyLabels = map[string]string{"openshift-cluster-group-upgrades/clusterGroupUpgrade": clusterGroupUpgrade.Name}
-	listOpts := []client.ListOption{
-		client.InNamespace(clusterGroupUpgrade.Namespace),
-		client.MatchingLabels(policyLabels),
-	}
-	policiesList := &unstructured.UnstructuredList{}
-	policiesList.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "policy.open-cluster-management.io",
-		Kind:    "PolicyList",
-		Version: "v1",
-	})
-	if err := r.List(ctx, policiesList, listOpts...); err != nil {
-		return nil, err
-	}
-
-	return policiesList, nil
 }
 
 func (r *ClusterGroupUpgradeReconciler) reconcileResources(ctx context.Context, clusterGroupUpgrade *ranv1alpha1.ClusterGroupUpgrade, managedPoliciesPresent []*unstructured.Unstructured) error {
@@ -2217,27 +2009,6 @@ func (r *ClusterGroupUpgradeReconciler) rootPolicyHandlerOnUpdate(ctx context.Co
 // SetupWithManager sets up the controller with the Manager.
 func (r *ClusterGroupUpgradeReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	r.Recorder = mgr.GetEventRecorderFor("ClusterGroupUpgrade")
-
-	placementRuleUnstructured := &unstructured.Unstructured{}
-	placementRuleUnstructured.SetGroupVersionKind(schema.GroupVersionKind{
-		Kind:    "PlacementRule",
-		Group:   "apps.open-cluster-management.io",
-		Version: "v1",
-	})
-
-	placementBindingUnstructured := &unstructured.Unstructured{}
-	placementBindingUnstructured.SetGroupVersionKind(schema.GroupVersionKind{
-		Kind:    "PlacementBinding",
-		Group:   "policy.open-cluster-management.io",
-		Version: "v1",
-	})
-
-	policyUnstructured := &unstructured.Unstructured{}
-	policyUnstructured.SetGroupVersionKind(schema.GroupVersionKind{
-		Kind:    "Policy",
-		Group:   "policy.open-cluster-management.io",
-		Version: "v1",
-	})
 
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&ranv1alpha1.ClusterGroupUpgrade{}, builder.WithPredicates(predicate.Funcs{
