+++ pkg/api/clustergroupupgrades/v1alpha1/types.go
@@ -250,8 +250,6 @@ type ClusterGroupUpgradeStatus struct {
 	PlacementBindings []string `json:"placementBindings,omitempty"`
 	// +operator-sdk:csv:customresourcedefinitions:type=status,displayName="Placement Rules"
 	PlacementRules []string `json:"placementRules,omitempty"`
-	// +operator-sdk:csv:customresourcedefinitions:type=status,displayName="Copied Policies"
-	CopiedPolicies []string `json:"copiedPolicies,omitempty"`
 	// +operator-sdk:csv:customresourcedefinitions:type=status,displayName="Conditions"
 	Conditions []metav1.Condition `json:"conditions,omitempty"`
 	// +operator-sdk:csv:customresourcedefinitions:type=status,displayName="Remediation Plan"
